+++ src/Client/Region/RegionErrorHandler.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CrazyGoat\TiKV\Client\RawKv;
+namespace CrazyGoat\TiKV\Client\Region;
 
 use CrazyGoat\TiKV\Client\Exception\RegionException;